Fixed circular import: bs.py doesnt ref qf.py
